requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
idna==3.4
joblib==1.2.0
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
praw==7.6.1
prawcore==2.3.0
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
tqdm==4.64.1
update-checker==0.18.0
urllib3==1.26.14
websocket-client==1.5.1
